update footer and logo
