+++ components/layout/select.tsx
@@ -2,7 +2,7 @@ import type { ReactElement, ReactNode } from "react";
 import cn from "clsx";
 import { Listbox, Transition } from "@headlessui/react";
 import { CheckIcon } from "nextra/icons";
-import { usePopper } from "../usePopper";
+import { usePopper } from "./usePopper";
 import { createPortal } from "react-dom";
 import { useMounted } from "nextra/hooks";
