+++ package.json
@@ -13,14 +13,17 @@
   },
   "homepage": "https://rooch.network",
   "dependencies": {
+    "@headlessui/react": "^1.7.14",
+    "@popperjs/core": "^2.11.7",
     "classnames": "2.3.2",
+    "clsx": "^1.2.1",
+    "cobe": "0.4.2",
     "next": "^13.2.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-spring": "^9.4.4",
     "nextra": "^2.2.20",
     "nextra-theme-docs": "^2.2.20",
-    "cobe": "0.4.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-spring": "^9.4.4"
   },
   "devDependencies": {
     "@types/node": "18.15.11",
